+++ src/libponyc/pass/pass.h
@@ -184,6 +184,7 @@ typedef struct pass_opt_t
   pass_id program_pass;
   bool release;
   bool library;
+  bool runtimebc;
   bool pic;
   bool ieee_math;
   bool print_stats;