+++ CHANGELOG.md
@@ -87,6 +87,7 @@ All notable changes to the Pony compiler and standard library will be documented
 - `DoNotOptimise` primitive
 - `compact` method for `Array` and `String`
 - `String.push_utf32()` method.
+- Allow the use of a LLVM bitcode file for the runtime instead of a static library.
 
 ### Changed